+++ changelog.md
@@ -24,3 +24,4 @@
 - make the contact us form workable by adding a URL to the request and changing the keys to match the received data
 # Changes in [1.0.5] (2023-06-25)
 - Integrated Arabic and English languages into the website
+- removed language flags from the header